Add cast
Former-commit-id: cb2aba11a153ff7fe66c7882482cfe109308fc39

+++ WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/src/vector_tiles/MapboxMultiSourceTileInfo.mm
@@ -88,7 +88,7 @@ public:
 
 - (int)maxZoom
 {
-    return sourcesByZoom.size()-1;
+    return (int)sourcesByZoom.size()-1;
 }
 
 - (bool)addImageMap:(NSString *)map minZoom:(int)minZoom maxZoom:(int)maxZoom type:(NSString *)imageType
@@ -99,7 +99,7 @@ public:
     source.minZoom = minZoom;  source.maxZoom = maxZoom;
     source.ext = imageType;
     sources.push_back(source);
-    [self addedSource:sources.size()-1];
+    [self addedSource:(int)sources.size()-1];
     
     return true;
 }
@@ -138,7 +138,7 @@ public:
     source.styleSet = styleSet;
     source.tileParser = tileParser;
     sources.push_back(source);
-    [self addedSource:sources.size()-1];
+    [self addedSource:(int)sources.size()-1];
     
     return true;
 }
@@ -174,7 +174,7 @@ public:
     else
         source.maxZoom = maxZoom;
     sources.push_back(source);
-    [self addedSource:sources.size()-1];
+    [self addedSource:(int)sources.size()-1];
     
     return true;
 }
